[AG-17118] Refactor(pagination): split page numbers into opt-in module
Move PageNumbersComp into its own PageNumbersModule, registered as a
dynamic bean so it tree-shakes out of PaginationModule when unused.
Extract page-number CSS and the page-item builder into dedicated files.
