+++ packages/ag-grid-community/src/context/context.ts
@@ -161,6 +161,7 @@ export type DynamicBeanName =
     | 'headerFilterCellCtrl'
     | 'headerGroupCellCtrl'
     | 'rangeHandle'
+    | 'pageNumbers'
     | 'tooltipFeature'
     | 'highlightTooltipFeature'
     | 'tooltipStateManager'
