+++ packages/ag-grid-community/src/interfaces/iModule.ts
@@ -165,6 +165,7 @@ export type CommunityModuleName =
     | 'NumberFilter'
     | 'BigIntFilter'
     | 'Pagination'
+    | 'PageNumbers'
     | 'PinnedRow'
     | 'QuickFilter'
     | 'RenderApi'
@@ -257,6 +258,7 @@ export type AgModuleName =
     | 'NumberFilterModule'
     | 'BigIntFilterModule'
     | 'PaginationModule'
+    | 'PageNumbersModule'
     | 'PinnedRowModule'
     | 'QuickFilterModule'
     | 'RenderApiModule'
