+++ packages/ag-grid-community/src/main.ts
@@ -1058,6 +1058,7 @@ export { LocaleModule } from './misc/locale/localeModule';
 export { GridStateModule } from './misc/state/stateModule';
 export { ModuleRegistry } from './modules/moduleRegistry';
 export { PaginationModule } from './pagination/paginationModule';
+export { PageNumbersModule } from './pagination/pageNumbersModule';
 export { PinnedRowModule } from './pinnedRowModel/pinnedRowModule';
 export { HighlightChangesModule } from './rendering/cell/highlightChangesModule';
 export { RenderApiModule } from './rendering/renderModule';
