+++ packages/ag-grid-community/src/pagination/paginationComp.css
@@ -41,48 +41,6 @@
     }
 }
 
-.ag-paging-page-numbers-panel {
-    display: flex;
-    align-items: center;
-    gap: calc(var(--ag-spacing) * 1.5);
-}
-
-.ag-paging-page-numbers {
-    display: flex;
-    align-items: center;
-    gap: calc(var(--ag-spacing) * 0.5);
-}
-
-.ag-paging-page-number {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    min-width: var(--ag-icon-size);
-    padding: 0 calc(var(--ag-spacing) * 0.5);
-    border-radius: var(--ag-border-radius);
-    cursor: pointer;
-    font-variant-numeric: tabular-nums;
-
-    &:hover {
-        background-color: var(--ag-row-hover-color);
-    }
-}
-
-.ag-paging-page-numbers-truncated .ag-paging-page-number {
-    padding: var(--ag-spacing);
-}
-
-.ag-paging-page-number-current {
-    cursor: default;
-    font-weight: 600;
-    background-color: color-mix(in srgb, var(--ag-background-color), var(--ag-foreground-color) 5%);
-}
-
-.ag-paging-page-number-ellipsis {
-    pointer-events: none;
-    font-variant-numeric: tabular-nums;
-}
-
 :where(.ag-number-field).ag-paging-number {
     display: inline-block;
 }
