+++ packages/ag-grid-community/src/validation/rules/dynamicBeanValidations.ts
@@ -9,6 +9,7 @@ export const DYNAMIC_BEAN_MODULES: Record<DynamicBeanName, ValidationModuleName>
     headerFilterCellCtrl: 'ColumnFilter',
     headerGroupCellCtrl: 'ColumnGroup',
     rangeHandle: 'CellSelection',
+    pageNumbers: 'PageNumbers',
     tooltipFeature: 'Tooltip',
     highlightTooltipFeature: 'Tooltip',
     tooltipStateManager: 'Tooltip',
